Refactor started
