+++ apps/dh/api-dh/source/DataHub.WebApi/Modules/MarketParticipant/User/Dto/GetUserResponse.cs
@@ -11,17 +11,7 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
+namespace Energinet.DataHub.WebApi.Clients.MarketParticipant.v1;
 
-using Energinet.DataHub.WebApi.GraphQL.Extensions;
-using MarketParticipant = Energinet.DataHub.WebApi.Clients.MarketParticipant.v1;
-
-namespace Energinet.DataHub.WebApi.GraphQL.Types.User;
-
-public class SortDirctionType : EnumType<MarketParticipant.SortDirection>
-{
-    protected override void Configure(IEnumTypeDescriptor<MarketParticipant.SortDirection> descriptor)
-    {
-        descriptor.Name("MarketParticipantSortDirctionType");
-        descriptor.AsIsCase();
-    }
-}
+public partial class GetUserResponse : IUser
+{ }
