+++ apps/dh/api-dh/source/DataHub.WebApi/Modules/MarketParticipant/User/Types/SortDirectionType.cs
@@ -12,19 +12,16 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using HotChocolate.Data.Sorting;
+using Energinet.DataHub.WebApi.Clients.MarketParticipant.v1;
+using Energinet.DataHub.WebApi.GraphQL.Extensions;
 
-namespace Energinet.DataHub.WebApi.GraphQL.Types.User;
+namespace Energinet.DataHub.WebApi.Modules.MarketParticipant.User.Types;
 
-public class UsersSortInputType : InputObjectType<UsersSortInput>
+public class SortDirctionType : EnumType<SortDirection>
 {
-    protected override void Configure(
-        IInputObjectTypeDescriptor<UsersSortInput> descriptor)
+    protected override void Configure(IEnumTypeDescriptor<SortDirection> descriptor)
     {
-        descriptor.Field(f => f.Name).Type<DefaultSortEnumType>();
-        descriptor.Field(f => f.Email).Type<DefaultSortEnumType>();
-        descriptor.Field(f => f.PhoneNumber).Type<DefaultSortEnumType>();
-        descriptor.Field(f => f.LatestLoginAt).Type<DefaultSortEnumType>();
-        descriptor.Field(f => f.Status).Type<DefaultSortEnumType>();
+        descriptor.Name("MarketParticipantSortDirctionType");
+        descriptor.AsIsCase();
     }
 }
