+++ apps/dh/api-dh/source/DataHub.WebApi/Modules/MarketParticipant/User/Types/UsersSortInput.cs
@@ -14,7 +14,7 @@
 
 using Energinet.DataHub.WebApi.Clients.MarketParticipant.v1;
 
-namespace Energinet.DataHub.WebApi.GraphQL.Types.User;
+namespace Energinet.DataHub.WebApi.Modules.MarketParticipant.User.Types;
 
 public record UsersSortInput(
     SortDirection? Name,
