+++ apps/dh/api-dh/source/DataHub.WebApi/Modules/MarketParticipant/User/Types/UsersSortInputType.cs
@@ -1,0 +1,30 @@
+// Copyright 2020 Energinet DataHub A/S
+//
+// Licensed under the Apache License, Version 2.0 (the "License2");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+using HotChocolate.Data.Sorting;
+
+namespace Energinet.DataHub.WebApi.Modules.MarketParticipant.User.Types;
+
+public class UsersSortInputType : InputObjectType<UsersSortInput>
+{
+    protected override void Configure(
+        IInputObjectTypeDescriptor<UsersSortInput> descriptor)
+    {
+        descriptor.Field(f => f.Name).Type<DefaultSortEnumType>();
+        descriptor.Field(f => f.Email).Type<DefaultSortEnumType>();
+        descriptor.Field(f => f.PhoneNumber).Type<DefaultSortEnumType>();
+        descriptor.Field(f => f.LatestLoginAt).Type<DefaultSortEnumType>();
+        descriptor.Field(f => f.Status).Type<DefaultSortEnumType>();
+    }
+}
